requirements.txt
fastapi
uvicorn
streamlit
requests
chromadb
sentence-transformers
openai
python-dotenv
